requirements.txt
ccxt
pandas
numpy
scikit-learn
requests
psycopg2-binary